fix: move the claude cloud spawner's deadline-racing test to the live tier
Tests/CLAUDE.md's test-tier taxonomy names this case verbatim: a suite
belongs in Tests/TBDDaemonLiveTests when it "spawns a child racing a
deadline." spawnMapsATimeoutToTimedOut raced a real 0.2s deadline
against a real child in the fast parallel pass, alongside
BoundedProcessRunnerTests (which covers the same runBoundedProcess
machinery but has zero timeout-racing tests) and every sibling
timeout suite already living in TBDDaemonLiveTests. Moved to a new
ClaudeCloudSpawnerTimeoutTests suite there and added a tier statement
to the suite doc comment in both files.
